Remove nullable Optional. All hail Optionull

+++ config-api/src/main/java/net/neoforged/camelot/api/config/storage/SQLStorage.java
@@ -1,9 +1,7 @@
 package net.neoforged.camelot.api.config.storage;
 
 import org.jdbi.v3.core.Jdbi;
-import org.jetbrains.annotations.Nullable;
 
-import java.util.Optional;
 import java.util.function.Function;
 
 class SQLStorage<G> implements ConfigStorage<G> {
@@ -18,34 +16,33 @@ class SQLStorage<G> implements ConfigStorage<G> {
     }
 
     @Override
-    public void restoreToDefault(String key, G target) {
-        jdbi.withHandle(handle -> handle.createUpdate("delete from " + tableName + " where target = ? and key = ?")
-                .bind(0, identifier.apply(target))
-                .bind(1, key)
-                .execute());
+    public void store(String key, G target, Optionull<String> value) {
+        if (value.isEmpty()) {
+            // Restore to default
+            jdbi.withHandle(handle -> handle.createUpdate("delete from " + tableName + " where target = ? and key = ?")
+                    .bind(0, identifier.apply(target))
+                    .bind(1, key)
+                    .execute());
+        } else {
+            jdbi.withHandle(handle -> handle.createUpdate("insert or replace into " + tableName + "(target, key, value) values (?, ?, ?)")
+                    .bind(0, identifier.apply(target))
+                    .bind(1, key)
+                    .bind(2, value.getValue())
+                    .execute());
+        }
     }
 
     @Override
-    public void store(String key, G target, String value) {
-        jdbi.withHandle(handle -> handle.createUpdate("insert or replace into " + tableName + "(target, key, value) values (?, ?, ?)")
-                .bind(0, identifier.apply(target))
-                .bind(1, key)
-                .bind(2, value)
-                .execute());
-    }
-
-    @Override
-    @Nullable
-    public Optional<String> read(String key, G target) {
+    public Optionull<String> read(String key, G target) {
         return jdbi.withHandle(handle -> handle.createQuery("select value from " + tableName + " where target = ? and key = ?")
                 .bind(0, identifier.apply(target))
                 .bind(1, key)
                 .execute((statementSupplier, ctx) -> {
                     var rs = statementSupplier.get().getResultSet();
                     if (rs.next()) {
-                        return Optional.ofNullable(rs.getString(1));
+                        return Optionull.of(rs.getString(1));
                     }
-                    return null;
+                    return Optionull.empty();
                 }));
     }
 }
